+++ test/runtests.jl
@@ -6,6 +6,7 @@ using TreeSitter, Test
     include("nodes.jl")
     include("queries.jl")
     include("predicates.jl")
+    include("introspection.jl")
     include("abstracttrees.jl")
     include("local_grammar.jl")
 end